[legacy] Fix RPATH information
With this commit the correct RPATH information is added to the shared
libraries and binaries of all installed packages. For packages not
adding the info already properly we add some CMake variables in our
meta build system to pass the needed information to the CMake based
build system of the individual packages.
With this change the FairSoft installtion is definitely not relocatable
any longer, but it is not clear if it was before.
With the changes the environment varaible (DY)LD_LIBRARY_PATH shouldn't
be needed any longer. But this fact still has to be tested.

This PR fixes https://github.com/FairRootGroup/FairSoft/issues/535